sched/tune: fix do_exit() accounting race condition
Due to the sequencing of schedtune_exit() and PF_EXITING flag it could
happen that a task already gone through schedtune_exit() is migrated
before it's marked PF_EXITING, thus resulting in being permanently
accounted in the last CPS has been enqueued after migration.

This patch move the schedtune_exit() call right after the task has been
marked PF_EXITING, thus granting the dequeue accounting from the CPU where
it's going to die.

Moreover, we export and make use of the core::task_rq_{un}lock() functions
to grant the locking of the correct RQ as well as serializing with the
migration code.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>