arm: Update SD_ASYM_CPUCAPACITY topology flag when cpu capacity is changed
When cpu capacity information is provided through DT or sysfs we might
not know if the system has asymmetric cpu capacities when the system
topology information is initially passed to the scheduler. We therefore
have to enforce a rebuild of the sched_domain hierarchy if we later
discover that the flag was initially set wrong.

It is mainly an issue for systems with same type of cores (dmips/mhz)
but different max frequencies, and asymmetric cpu capacity systems where
cpu capacity is not provided by DT but set through sysfs later.

cc: Russell King <linux@armlinux.org.uk>

Signed-off-by: Morten Rasmussen <morten.rasmussen@arm.com>