sched/fair: add path to migrate to higher capacity CPU
In old code there has only one possible to directly migrate task for lower
capacity CPU to higher capacity CPU: the previous CPU's capacity cannot
meet the bandwidth requirement and the CPU is not over-utilized. This is
reasonable if we only think to task util_avg signal to judge the task
performance requirement with comparing with CPU capacity.

But in some scenarios, the task takes much time to stay on runnable
state, so its load_avg value is quite higher than util_avg. So for this
case can optimize to use load_avg to help migration decision.

If so, we need add path to migrate to higher capacity CPU if destination
CPU has higher capacity than lower capacity CPU.

Signed-off-by: Leo Yan <leo.yan@linaro.org>