save current status
CFS throttled task list in thottled group's rt_rq doesn't seem to be
really feasible, or it's at least quite hard to implement.

Main stopper is probably the necessity to grab both pi_lock and rq_lock
(in this order), which is doable in the throttling path but seems hard
for unthrottling.

Current code also uses rt_nr_running for knowing what to do and that is
updated if we remove CFS throttled tasks from group's rt_rq. This can
probably be fixed with a rt_nr_cfs_throttled counter anyway.