schedtune: Add support for negative boost values
Change-Id: I164ee04ba98c3a776605f18cb65ee61b3e917939

sched/fair: Using boosted_task_util to calculate new_util in energy_aware_wake_cpu.

Improves performance of vellamo suite by ~8% to 10% on Sailfish.

Change-Id: I78f957ae39507a451ac3437200771391269630bc

sched/fair: Moving boosted_task_util outside the loop.

Change-Id: I79c4ab2037ff03b7c99625c3655532786d3aac56

schedtune: handle accounting error in schedtune boost groups

This works around the case where a task is enqueued while a member
of one sched-tune group, but the task is migrated to another group
before being dequeued. This would manifest as an attempt to
decrement below zero which would be ignored -- but the accounting
for the old boost group would never be decremented resulting in
CPU boosts that would never return to zero (possibly causing
higher frequencies).

A proper fix is coming with a callback that adjusts the accounting
when the task is moved to a different group.

Change-Id: I18be893c4433d7fd3b3b23da7c734f37acda3c43

sched/tune: fix payoff calculation for boost region

The definition of the acceptance regions as well as the translation of
these regions into a payoff value was both wrong which turned out in:
a) a wrong definition of payoff for the performance boost region
b) a correct "by chance" definition of the payoff for the performance
   constraint region (i.e. two sign errors together fixing the formula)

This patch provides a better description of the cut regions as well as
a fixed version of the payoff computations, which are now reduced to a
single formula usable for both cases.

Reported-by: Leo Yan <leo.yan@linaro.org>
Reviewed-by: Leo Yan <leo.yan@linaro.org>
Signed-off-by: Leo Yan <leo.yan@linaro.org>
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>

sched/tune: fix compilation error on !CONFIG_CGROUP_SCHEDTUNE

When SchedTune is configured without CGroups support, the schedtune_accept_deltas()
function is not declared.
Since schedule candidate filtering is required also when global boosting
mode is in use, this patch move the declaration of that function to make
it dependant just on CONFIG_SCHED_TUNE.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>

sched/{fair,tune}: simplify fair.c code

The usage of conditional compiled code is discouraged in fair.c.

This patch clean up a bit fair.c by moving schedtune_{cpu.task}_boost
definitions into tune.h.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>

sched/tune: use a single initialisation function

With the introduction of initialization function required to compute the
energy normalization constants from DTB at boot time, we have now a
late_initcall which is already used by SchedTune.

This patch consolidate within that function the other initialization
bits which was previously deferred to the first CGroup creation.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>

sched/tune: beautify schedtune_tasks_update()

The schedtune_tasks_update() function is used to update the accouting
of tasks on a certain boost group. It support only add/remove operations
and provide proper trace events as well as CPU boost update if
required.

This patch does not change the funtional bits but contributs to cleanin
up a bit the code to make it more simple and readable from call sites.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>

Conflicts:
	kernel/sched/tune.c

sched/tune: fix accounting for runnable tasks (1/5)

The accounting for tasks into boost groups of different CPUs is currently
broken mainly because:
a) we do not properly track the change of boost group of a RUNNABLE task
b) there are race conditions between migration code and acconting code

This patch provides a first chunk of fixes to ensure enqueue/dequeue
accounting also for throttled tasks.

Without this patch is can happen that a task is enqueued into a throttled
RQ thus not being accounted for the boosting of the corresponding RQ.
We could argue that a throttled task should not boost a CPU, however:
a) properly implementing CPU boosting considering throttled tasks will
   increase a lot the complexity of the solution
b) it's not easy to quantify the benefits introduced by such a more
   complex solution

Since task throttling requires the usage of the CFS bandwidth controller,
which is not widely used on mobile systems (at least not by Android kernels
so far), for the time being we go for the simple solution and boost also
for throttled RQs.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>

sched/tune: fix accounting for runnable tasks (2/5)

The accounting for tasks into boost groups of different CPUs is currently
broken mainly because:
a) we do not properly track the change of boost group of a RUNNABLE task
b) there are race conditions between migration code and acconting code

This patch provides the code required to enforce proper locking.
A per boost group spinlock has been added to grant atomic
accounting of tasks as well as to serialise enqueue/dequeue operations,
triggered by tasks migrations, with cgroups's attach/detach operations.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>

sched/tune: fix accounting for runnable tasks (3/5)

The accounting for tasks into boost groups of different CPUs is currently
broken mainly because:
a) we do not properly track the change of boost group of a RUNNABLE task
b) there are race conditions between migration code and acconting code

This patch adds cgroups {allow,can,cancel}_attach callbacks.

Since a task can be migrated between boost groups while it's running,
the CGroups's attach callbacks have been added to properly migrate
boost contributions of RUNNABLE tasks.

The RQ's lock is used to serialise enqueue/dequeue operations, triggered
by tasks migrations, with cgroups's attach/detach operations. While the
SchedTune's CPU lock is used to grant atrocity of the accounting within
the CPU.

NOTE: the current implementation does not allows a concurrent CPU migration
      and CGroups change.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>

Conflicts:
	kernel/sched/tune.c

sched/tune: fix accounting for runnable tasks (4/5)

The accounting for tasks into boost groups of different CPUs is currently
broken mainly because:
a) we do not properly track the change of boost group of a RUNNABLE task
b) there are race conditions between migration code and acconting code

This fixes accounting for exiting tasks by adding a dedicated call early
in the do_exit() syscall, which disables SchedTune accounting as soon as a
task is flagged PF_EXITING.

This flag is set before the multiple dequeue/enqueue dance triggered
by cgroup_exit() which is useful only to inject useless tasks movements
thus increasing possibilities for race conditions with the migration code.
The schedtune_exit_task() call does the last dequeue of a task from its
current boost group. This is a solution more aligned with what happens in
mainline kernels (>v4.4) where the exit_cgroup does not move anymore a dying
task to the root control group.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>

sched/tune: fix accounting for runnable tasks (5/5)

The accounting for tasks into boost groups of different CPUs is currently
broken mainly because:
a) we do not properly track the change of boost group of a RUNNABLE task
b) there are race conditions between migration code and accounting code

To avoid accounting issues at startup, this patch disable the SchedTune
accounting until the required data structures have been properly
initialized.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>