sched: fix SchedFreq integration for both PELT and WALT
The current kernel allows to use either PELT or WALT to track CPUs utilizations.
One of the main differences between the two approaches is that PELT
tracks only utilization of SCHED_OTHER classes while WALT tracks all tasks
with a single signal.

The current sched_freq_tick does not make this distinction and, when WALT
is in use, we end up adding multiple time the contribution related to
the RT and DL classes. This patch fixes this issue by:

1. providing two different code paths for PELT and WALT, thus granting that
   when we switch to PELT we get the original behaviour based on the assumption
   that class aggregations is done underneath by SchedFreq.

2. avoiding the double accounting of DL and RT workloads, when WALT is in use,
   by just adding a margin to the original WALT signal when we need to check
   if the CFS capacity has to be increased.

Change-Id: I7326fd50e868e97fb5e12351917e9d2969bfdae7
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>

Conflicts:
	kernel/sched/core.c

sched: fix set_cfs_cpu_capacity when WALT is in use
sched: fix SchedFreq integration for both PELT and WALT

This reverts commit c0cc03e1e8423787dc363ec0ffda5afc78e8f7af.
Originally reverted as it caused perf regression in camera. However,
with ag/1203633 (that puts cameraserver in foreground schedtune cgroup),
the regression went away.

Ran all tests in go/camera-performance-testing-checklist with and
without this CL and saw no perf regression. Saw significant power
savings ranging from 5% to 20%.

BUG: 28312446
Change-Id: Ia178f81c6ae29ebd21e480cca27a751f32d74927