sched/fair: add support for performance index
SchedTune allows to evaluate and enfore a trade-off between energy
saving and performance boosting. This requries to defined a performance
metrics which allows to esitmate what are the impact on performances for
a speficic scheduling decision.

This patch introduces in the new energy_diff function the required
computation of a new PerformanceIndex, which is considers both:
- the SpeedUp of a task running on an higher capacity CPU
- the Delay of a task being co-scheduled with others

This is just an initial proposal for the performance index which can be
extended in the future to account for example for interrupts,
utilization generated by higher priority classes, etc.

cc: Morten Rasmussen <morten.rasmussen@arm.com>
cc: Chris Redpath <Chris.Redpath@arm.com>
cc: Juri Lelli <Juri.Lelli@arm.com>
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>