sched/trace: Add task migration cause to trace_sched_migrate_task
Enhance the trace event 'sched_migrate_task' for CFS tasks by providing
the actual cause of the task migration.

This is useful in case you want to evaluate the migration history of a
specific task of a workload.

This feature can be enabled by setting CONFIG_SCHED_DEBUG_EAS_MIGRATION
(disabled by default).

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>