sched/fair: keep boostgroups into consideration in find_new_capacity()
In find_new_capacity() we look for an OPP which has enough capacity for
a specified task. However we are not considering the boost value for that
task.

This patch ensure that the capacity returned is considering the boost
value of the task and returns the maximum between the cluster capacity
and the capacity required to accommodate the task boosted utilisation.

NOTE: this implementation can be improved by caching capacity index once
computed for a certain frequency domain. However, such an optiomization
is more complex in the bottom-up approach used by the current
energy_diff implementation.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>