sched/events: Introduce task group PELT trace event sched_pelt_tg()
The trace event is only defined if symmetric multi-processing
(CONFIG_SMP) and group scheduling support for SCHED_OTHER
(CONFIG_FAIR_GROUP_SCHED) are enabled.

The cfs_rq owned by the task group is used as the only parameter for the
trace event because it has a reference to the task group and the cpu.
Using the task group as a parameter instead would require the cpu as a
second parameter.

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>