sched/fair: consider task's boost in find_new_capacity()
In systems where multiple CPUs shares the same frequency domain, a
boosted task on a CPU bias the minimum capacity for all the CPUs in the
same SG. Currently, while computing the capacity of a SG we do not
consider the minimum capacity required by a boosted tasks.

Make sure that the minimum capacity reported by find_min_capacity
is compliant with the minimum capacity required by the current task.
Indeed, even if the task should remain alone in that SG, the minimum
capacity required by its boosting value should be granted.

Change-Id: Ia89da70ff9c0336d53b77e78898b7f67a5992be7
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Signed-off-by: Chris Redpath <chris.redpath@arm.com>