Experimental!: sched/fair: Code !is_big_little code path into select_energy_cpu_brute()
This patch creates a synch test point so that we can see how the new
wakeup path behaves on the Pixel device.

Future work is needed to intergate find_best_target() in a better way. The
aim is to support !is_big_little and is_big_little systems without the
exisiting sys-fs switch.

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>