sched/fair: Fix PELT time-stamp precision
PELT is fed with nanosecond precision time-stamps, but only uses
micro-second precision for all calculations. The time delta since last
update is currently calculated in nano-second precision and then shifted
to get the micro-second delta meaning that sub-micro-second chunks of
time are lost and not accounted for.

To resolve this issue __update_load_avg() must either be fed with
micro-second precision time-stamps instead, or the time-stamps must be
round down internally. The first option requires more changes and makes
__update_load_avg() special in that time-stamps are generally
nano-second precision in the scheduler. This patch therefore implements
the latter approach.