sched/core: add capacity_{min,max} to CPU controller
The CPU CGroup controller allows to assign a specified (maximum)
bandwidth to tasks within a group, however it does not enforce any
constraint on how such bandwidth can be consumed.
With the integration of schedutil, the scheduler has now the proper
information about a task to select  the most suitable frequency to
satisfy tasks needs.

This patch extends the CPU controller by adding a couple of new
attributes, capacity_min and capacity_max, which can be used to enforce
bandwidth boosting and capping. More specifically:

- capacity_min: defines the minimum capacity which should be granted
                (by schedutil) when a task in this group is running

- capacity_max: defines the maximum capacity which can be granted
                (by schedutil) when a task in this group is running

These attributes are:
a) tunable at all hierarchy levels, i.e. root group as well
b) allows to create subgroups of tasks which are not violating the
   capacity constraints defined by the father group.
   Thus, tasks on a subgroup can only be more boosted and/or more
   capped, which is matching with the "limits" schema proposed by
   the "Resource Distribution Model (RDM)" suggested by the
   CGroups v2 documentation.

This patch provides the basic support to expose the two new attributes
and to validate their run-time update based on the "limits" schema
of the RDM.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>