sched/core: track CPU's capacity_{min,max}
At run-time, the capacity constraints of RUNNABLE tasks must be
aggregated to figure out what are the capacity constraints to enforce
on each CPU.

This aggregation must meet two main goals:
  1) ensure the minimum capacity required by the most boosted
     RUNNABLE task on that CPU
  2) do not penalize the less capped RUNNABLE task

Thus, the aggregation for both the capacity constraints turns out to be
a MAX function on the min/max capacities of RUNNABLE tasks.

  cpu_capacity_min := MAX(capacity_min_i), for each RUNNABLE task_i
  cpu_capacity_max := MAX(capacity_max_i), for each RUNNABLE task_i

Since the CPU controller uses a SLAB allocator for the allocation of
task_group structures, the proposed approach adds just a couple of
unsigned int to the task_group, thus adding negligible memory overheads.

The aggregation at CPU level is done by exploiting the task_struct.
Tasks are already enqueued, via their embedded scheduling entities, in
many different lists and trees. This patch uses the same approach to
keep track of the capacity constraints enforced by every task on a CPU.
To this purpose:
  - each CPU has two RBTree, for the tracking of minimum and maximum
    capacity constraints all the tasks enqueue on that CPU
  - task_struct has two rb_node which allows to position that task in the
    minimum/maximum capacity tracking RBTree of the CPU in which the
    task is enqueued

This patch provides the RBTree support code while, for the sake of
clarity, the synchronization between the
   fast path: {enqueue,dequeue}_task
and the
   slow path: cpu_capacity_{min,max}_write_u64
is provided in a dedicated path.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>