sched/core: sync capacity_{min,max} between slow and fast paths
At enqueue/dequeue time a task needs to be placed in the CPU's rb_tree
depending on the current capacity_{min,max} value of the CPU cgroup it
belongs to. Thus, we need to grant that these values cannot be changed
while the task is in these critical sections.

To this purpose, this patch adds a dedicated "cap_lock" spinlock to the
task_struct which protects the updates of capacity_{min,max} values for
RUNNABLE tasks.

This patch adds the synchronization calls required to grant that each
RUNNABLE task is always in the correct relative position within the
RBTree. Specifically, when a group's capacity_{min,max} value is
updated, all the currently RUNNABLE tasks are re-positioned within the
rb_tree, if required by a change in their relative position.
This operation is mutually exclusive with the task being {en,de}queued
thanks to the usage of the cap_lock.

It's worth to notice that moving a task from a CGroup to another,
perhaps with different capacity_{min,max} values, is already covered by
the current locking schema. Indeed, this operation requires a dequeue
from the original RQ followed by an enqueue in the new one.
The same argument is true for tasks migrations thus, tasks migrations
between CPUs and CGruoups are ultimately managed like tasks
wakeups/sleeps.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>