sched/fair: Task PELT idle decay clamping
Periodic tasks with long periods loose most of their accumulated
load/utilization between each activation causing any decisions about
compute capacity to be off by a significant margin. A potential solution
to the problem is to 'clamp' the decay during and effectively make
idle-time beyond a set threshold disappear so the task will appear as
having a shorter period and therefore have a higher load/utilization
when it wakes up.

The main implementation issue with this idea is that aggregate PELT
signals for cfs_rqs and group entities have to be updated with the
reduced decay load/utilization of the task every time decay clamping
happens at task wake-up. Currently we only have to add/subtract
load/utilization when tasks migrate.

In this patch decay clamping is accounted for by adding the delta
between the full decay and the clamped decay to the cfs_rq PELT signal.

Suggested-by: Paul Turner <pjt@google.com>
Suggested-by: Peter Zijlstra <peterz@infradead.org>