sched/fair: Experimental tracking of waiting time for decay clamping
One issue with time-based decay clamping is that co-scheduling of
periodic tasks can have a significant effect on the clamp for clamp
periods which are not significantly longer than the schedule period
(sched_latency). Not only will preemption cause tasks to reach a lower
maximum utilization, it will also decay more as utilization decay during
waiting time does not affect the decay clamp. So preempted periodic
tasks end up decaying their utilization for waiting_time + decay_clamp
time per task period, while tasks which are not preempted only decay for
decay_clamp time.

One way of mitigating this issue is to track waiting time and subtract
it from the clamp when calculating the utilization decay. This patch is
a _hack_ that implements this idea. Waiting time is subtracted from the
decay of load as well, which is wrong. It can be done it is requires
refactoring of __update_load_avg().

This is purely for experimentation.